Added ImageUser class; re-factored Invader Class.
-Added an ImageUser abstract class to house some of the methods and
fields/properties the PlayerShip and Invader classes once shared, such
as Location, Area, and ResizeImage. Both the PlayerShip class and the
Invader classes now inherit from this ImageUser class for those methods.

-Added a bitmap array to the Invader class to be loaded with the resized
images at the start. Invader's Draw() method will now use this array.
Also added methods to load Bug, Satellite, Star, Spaceship, and Saucer
images into this bitmap array, and a method to decide which images to
load based on invader ship type. This approach replaces the old method
of deciding which image to select and resizing it on the fly for every
single frame.