Added a Mothership.
-Added a Mothership that appears every round. It will appear near the
top of the screen, and go from left to right one time (unlike the rest
of the invaders, which switch when near the border). It appears and
disappears with an explosion. I added methods to the Game class to
supplement the Mothership. Along this line, I also made Invader an
abstract class that has MotherShip and InvaderGrunt as subclasses. Now
the grunts (regular invaders) and the mothership can have slightly
different behaviors.

-In terms of plain refactoring, I split the ImageUser class once again.
Graphic objects that are custom and those that have pictures backing
them both use a rectangular area which will have points that need
accessing. I made an AreaUser class which ImageUser now inherits from
for its Area-related functions.

Note: all these changes are fresh, so there might be some things I
forgot to comment.